[Port 2021.3][BuiltTarget Extensions]API was accessing the target plaform by the wrong string
UI is adding included/excluded platforms in the UI by NamedBuildTarget.

So for instance, "Switch" is obtained from BuildTarget.Switch.ToString(), but if we do NamedBuildTarget(BuildTarget.Switch) it returns you "Nintendo Switch".

This was including/excluding RP assets that were not supposed to match the call to the API of BuiltTarget extension

https://jira.unity3d.com/browse/UUM-73363