Backport wireframe fixes (including Vulkan)
Manual backport of [this PR](https://github.cds.internal.unity3d.com/unity/unity/pull/50412) + [this PR](https://github.cds.internal.unity3d.com/unity/unity/pull/25910).

Tested on D3D11 and Vulkan. Works well, the wireframe is now rendered.
It also fixes an issue with the DepthCopy, that is happening since 21.3 is not using the Blitter code introduced in 2022 + the changes for this pass.

![Fix_21_3_Wireframe](https://media.github.cds.internal.unity3d.com/user/1911/files/80ebf77a-4dee-4bec-96b5-e63032e60b1c)