[VFX] Fix random texture rendered with instancing and exposed texture set to None
If there are several instances of a VFX that have a exposed texture used in the Output context, and one of them is overriding the exposed texture value with an empty texture (None), this VFX will not render with the expected texture, and will use any of the textures used by the other instances.
This occurs because NULL textures are ignored when set directly during rendering.
This PR avoid using NULL textures and selects the correct default texture early.
The expected texture color should be:
- Using Shader Graph: Defined by the default value of the exposed texture in SG (White, Black, Blue...)
- Not using Shader Graph: Gray 50%

The issue is easy to reproduce, just create 2 or more instances of a VFX using a exposed texture in an Output context and override the exposed texture in one of them with None. More than one VFX should be visible at the same time for the issue to reproduce.
Depending on the version, the issue could be a random flickering texture or a stable wrong texture. This fix works in both cases.