[Port] [2022.3] Make ProbeSettings.cubeResolution public
Fix [this issue](https://jira.unity3d.com/browse/UUM-74949) raised by this [forum discussion](https://forum.unity.com/threads/solved-set-reflection-probe-resolution-in-script.1606875/).
TLDR:
ProbeSettings.cubeResolution was internal, making it impossible to change the resolution of a cube reflection probe at runtime, unless using Reflection or ASMRef.
The use case for this is to change the resolution of a realtime or on demand probe depending on it's distance to the camera, which is quite a valid use to save a bit of performances.