Make lightmaps and reflection probes respect Preset Manager (case UUM-70818)
As explained in [UUM-70818](https://jira.unity3d.com/browse/UUM-70818), presets currently do not apply to any reflection probes and lightmaps which Unity generate. This PR fixes that.

Note that an important consequence of this change is that any asset presets which the user have in their projects, will be applied in their entirety. This include weird setting, such as marking lightmap as a cookie. The rationale behind this is that if the user explicitly tells us to do X (by setting up a preset), then we _should_ do X. Of course, the user can use preset filters to stop their presets from hitting lightmaps and then can use the Exclude Property feature to only having a subset of properties being applied to baked assets (lightmaps + reflection probes).