[2022.3][VFX][Fix] Fix Screen Space Size (and other) block in ShaderGraph outputs
This PR fixes the blocks that are implicitly using TransformObjectToWorld (e.g. Screen Space Size) in a ShaderGraph output.
Prior to this PR, the block was trying to use `UNITY_MATRIX_M `which is redefined as `elementToWorld` ("particle -> world matrix), instead of the "object -> world" matrix. It now uses the correct matrix.
Additionally, the selection & picking passes were broken when using blocks like Screen Space Size so it required change.