[Submit Render Request]Allow nesting a Submit Render Request from RenderPipelineManager callbacks.
In cpp all the code paths to `ExtractAndExecuteRenderPipeline`:

![imagen](https://media.github.cds.internal.unity3d.com/user/3279/files/d22ea4a2-879b-482d-80b1-613de66e0447)

In all of the places, we create a local `ScriptableRenderContext`

```
 ScriptableRenderContext renderContext;
 renderContext.ExtractAndExecuteRenderPipeline(cameras, ...);
```

As we had a static boolean, to avoid recursion , it was imposible to execute another camera render when any camera in SRP has started rendering. That make imposible nesting `SubmitRenderRequest` and subscribe to `RenderPipelineManager.beginContextRendering or RenderPipelineManager.beginCameraRendering`.

If we remove the static variable and we allow having an scenario like:

```
Camera cameraA; //our main camera that renders the scene
Camera cameraB; //camera only rendered with RenderRequest

void OnCameraRender(ScriptableRenderContext ctx, Camera camera)
{
	RenderPipeline.SubmitRenderRequest(cameraB, request);
}

void Start()
{
//this adds a callback to render cameraB when a camera is rendered
RenderPipelineManager.beginCameraRendering += OnCameraRender;	
}
```

![imagen](https://media.github.cds.internal.unity3d.com/user/3279/files/f7ba90f0-cd63-44ad-9109-657e8e6303c3)

The previous static variable was introduced to avoid this kind of scenario where the user can do recursive rendering as CameraB when rendering will call the same `RenderPipelineManager `events.

Now what we avoid is that the same camera executes a recursive rendering. Nofiying the user. and he can avoid doing a recursive loop by checking if the `beginCameraRendering `was being done by `CameraA`, then execute the `SubmitRenderRequest `of `CameraB`.

If the users do so, we show them:
![imagen](https://media.github.cds.internal.unity3d.com/user/3279/files/6e2d56f7-a082-428e-98ec-687c2bb62294)