Move test volume component to runtime assemblies
Fix [UUM-76007](https://jira.unity3d.com/browse/UUM-76007). In some tests, some volume component got defined in Editor assembly instead of Runtime one. It was silent for long time as the tests was editor only. It appears when user try to install core package locally by copying it into the Package folder.

This PR do not cover the fact that Test folder is not discarded when a package is added in such a way (It seems not the case when locally added with "local:<path>" in the Package.json). But now we are not crashing the build anymore. Though User will have unwanted additional test volume if proceeding this way. Luckilly it is easy to remove the test folder at this point.