[UUM-72143] Fix URP Lens Flare with scaled pixelRect
Jira: [UUM-72143](https://jira.unity3d.com/browse/UUM-72143)

As reported in JIRA, URP LensFlare does not render correctly when Resolution Scaling is non-1.0. In the JIRA description, it says it renders incorrectly when RenderGraph is enabled, but it renders correctly as expected. However, it does not render as expected when RenderGraph is disabled because the deprecated original render pass uses an unscaled pixel rect.
By fixing the LensFlare viewport size to support the scaled pixel rect, it renders as expected when RenderGraph is disabled.