change material type from lit to screenspace
This PR fixes a bug in the TrailEffect sample when the trail effect wasn't being drawn.

Jira: https://jira.unity3d.com/browse/UUM-73668

Shader graph that was used in a custom pass to restore a copy of the previous frame was incorrectly using the 'Lit' material type:

![image](https://media.github.cds.internal.unity3d.com/user/760/files/b0edda7f-0df6-4aa4-922b-23fad11919e4)

where as the C# script was using the DrawProcedural API

https://docs.unity3d.com/6000.0/Documentation/ScriptReference/Rendering.CommandBuffer.DrawProcedural.html


that expects a 'Fullscreen' material type.

![image](https://media.github.cds.internal.unity3d.com/user/760/files/0c1c9ff1-deb7-49a3-8f8e-b4e2351a8e18)

This PR changes the material type from 'Lit' to 'Fullscreen' which results in the correct application of previous frame's copy.