Fix for glClientWaitSync: Expected application to have kicked everyth…
Issue : https://jira.unity3d.com/browse/UUM-63146
This PR tries to address an issue where in **glClientWaitSync: Expected application to have kicked everything until job: 96089 (possibly by calling glFlush)"** messages are spammed in the log and game crashes on some devices with PowerVR Rogue GE8320 GPU. This issue was happening with URP and not with HDRP or built-in render pipeline.

We started investigating the passes in URP and have narrowed down to MainLightShadowCasterPass, AdditionalLightShadowCasterPass having something related to the issue. When we are detecting we have an emptyshadowmap that needs to be rendered, we are returning without clearing the current render target. When I added a code to ResetTarget in such scenarios, to both the passes,  the issue seems to be resolved.