[UUM-70119] Fix render texture corruption when using HDR
UUM-70926

With HDR disabled, the UI is rendered only from the main camera. With HDR enabled, UI rendering occurs from all cameras.

RawImage can use a render texture as a source. Each camera that outputs to a render texture references the unprocessed texture when rendering the UI. This causes corrupted rendering.

Changed the UI to be rendered only on the main camera.