[SG] Sample Bug Fixes
Addresses UUM-76505 by updating the instructions for increasing the variant count limit to include instructions for re-importing the Lit shader after increasing the variant count.
Addresses UUM-76080 by updating all of the files in the Node Reference sample so they're compatible with Unity 6. I also did the same thing with the files in the Procedural Patterns sample.