Render Graph - Fix error when a pass is culled
Let's consider a simple scenario, a render graph with 3 passes:
- `myPass1` (index 0 in the contextData.passData `NativeList`)
- `myPass2` (index 1 in the contextData.passData `NativeList`)
- `myPass3`  (index 2 in the contextData.passData `NativeList`)

Until now, when merging multiple passes together (so the 3 previous passes into 1 NativePassData; assuming the merge can be done), we were not considering the fact that a pass can be culled.

So without any pass being culled, when merging the passes together, it gives a `NativePassData`, with the value:
```
myNativePass.firstGraphPass = 0; (index 0, representing `myPass1`)
myNativePass.numGraphPasses = 3; (since we are merging 3 passes together)

// You can retrieve the last index by doing:
var lastGraphPass = myNativePass.firstGraphPass + myNativePass.numGraphPasses - 1;
// So it gives lastGraphPass = 0 + 3 - 1 = 2 (index 2, representing `myPass3`)
```
So when iterating through all the passes, like this:
```
for (int i = 0; i < nativePass.numGraphPasses - 1; i++)
```

It works well when we don't have any culled passes, since for this scenario, `numGraphPasses` is equal to 3 as expected.

Now let's assume `myPass2` has been culled; the previous `for` doesn't work anymore, since `numGraphPasses` is now equal to 2. 
So we will iterate the first 2 passes of `contextData.passData`, so `myPass1` and `myPass2`, but it's wrong. We should iterate over `myPass1` and `myPass3`, skipping `myPass2` since it has been culled.

This wrong logic creates many bugs in the core RenderGraph system and the viewer.

Here's how this PR fixes this bug. 
Now, in addition to storing the index of the first pass (`nativePass.firstGraphPass`) and the number of passes (`nativePass.numGraphPasses`), we also store the index of the last pass (`nativePass.lastGraphPass`). 

So we can now calculate the `real count` of passes, including the culled ones, by doing :
```
var countPasses = nativePass.lastGraphPass - nativePass.firstGraphPass + 1;
```

Which gives, for our previous example (myPass2 being culled):
```
var countPasses = myPass3 (index 2) - myPass1 (index 0) + 1;
var countPasses = 2 - 1 + 1 = 3 passes
// This is now correct, it includes the culled pass. If we were using nativePass.numGraphPasse, it would gives 2, which is wrong since myPass2 has been culled.
```

With this fix, when iterating through passes, we simply check if the pass has been culled or not, avoiding using it or updating it as we were doing before. And we are now updating the non-culled pass, since now we iterate through them.

Here's a gif with the problem in action, before resolving it with this PR:

![WeirdMerge1](https://media.github.cds.internal.unity3d.com/user/1911/files/5a461c19-286d-4df2-8495-ecd480763c7a)