Fix the urp rendergraph full screen render feature.
Fix UUM-74448
The RenderGraph implementation of the full screen render feature is inefficient with memory.

It uses all global textures, which is unneeded in most if not all cases, increasing memory usage because the lifetime of the resources is longer than needed.

This PR uses the input requirements set on the feature to set what global is used. 

This can be a significant memory improvement/optimization when using the full screen render feature under certain circumstances.

Ideally, we avoid the copy when fetching the current camera color. We can do this by swapping the camera color resource. However, we would need to be sure that the new resource is completely overwritten, not only partially which requires the copy. That requires user input, ie, new UI. Additionally, we could optimize the copy using framebuffer fetch using RenderGraph.AddCopyPass utility. This is out of scope for this PR.

This PR also improves the 176_FullScreenPass_MultiPass_StencilTest automated tests. The green checkerboard pass doesn't require the color fetch so it's turned off now. This ensures we also test that code path, increasing coverage. 


before
![image](https://media.github.cds.internal.unity3d.com/user/1031/files/f25b8da4-c960-423d-8328-ca5d266cf57c)

after
![image](https://media.github.cds.internal.unity3d.com/user/1031/files/23ad240c-dd64-4aa7-9539-eb242d1902ab)