Disable depth priming when MSAA is enabled
This change removes the option to use Depth Priming in URP when MSAA is also enabled. If MSAA is enabled, a new GUI warning box will appear when trying to enabled Depth Priming, informing the user that this configuration is not possible.

The reason for this change is that Depth Priming with MSAA is not fully implemented and causes several bugs, but it is not performant either, which undermines the point of using Depth Priming in the first place.

![image](https://media.github.cds.internal.unity3d.com/user/7119/files/375d0bfc-b462-4feb-ab38-583c6b74863f)

This fixes multiple bugs:
https://jira.unity3d.com/browse/UUM-59480
https://jira.unity3d.com/browse/UUM-65800
https://jira.unity3d.com/browse/UUM-72556

Additionally, a few URP Depth Priming tests have been updated to ensure that they do not use MSAA so they remain functional. One test has been disabled for XR because that configuration forces MSAA.

Docs ticket: https://jira.unity3d.com/browse/DOCG-5943