[UUM-76570] Fix color picker causing main preview to stop updating
This PR fixes https://jira.unity3d.com/browse/UUM-75670.

When an auxiliary window like a color picker shows up, it takes focus away from the SG window. This causes the main preview to stop updating, and does not let the user see the color picker result until it is closed.

By instead using visibility (determined by the parent window's BecameVisible/BecameInvisible events) as the determiner, we can display the result but also stop processing when the window is hidden.