Improve URP RenderGraph sample code - Remove copy from custom post processing feature template
URP RenderGraph [Jira](https://jira.unity3d.com/browse/URP-1297).
This is documentation in the form of sample code in the asset template.

This PR provides more performant sample code that is part of the template to generate a new URP Post Processing Effect (with Volume).

This change removes the cameraColor copy from the template. This code is meant as a sample and a starting point for users to implement their own. 

The assumption is that this is the most common usage case, the full screen (every pixel) is written to, so we can save the copy and use the new ContextContainer resource swapping.

![image](https://media.github.cds.internal.unity3d.com/user/1031/files/fd65699e-01c7-460f-8a3a-09acd275d08b)

After (no copy, new resource is used by next pass)
![image](https://media.github.cds.internal.unity3d.com/user/1031/files/e4b8a732-7881-4913-aa17-d1b3cf311881)