[Port] [2021.3] [UUM-76270] Fix resize and scrolling behavior of Custom Function node's body field
This PR fixes https://jira.unity3d.com/browse/UUM-76270.

The Custom Function node's body field was difficult to use due to a combination of broken sizing and a lack of scrollbars. This fix prevents the body field from growing beyond the inspector width, ~~and enables horizontal/vertical scrollbars~~. **Note: TextField does not have scrolling support in 2021.3, so the scrollbar addition was not backported. Other ports are not affected**.

![image](https://media.github.cds.internal.unity3d.com/user/3965/files/053d0c59-a251-4484-b532-f72ab0f37eaf)