[UUM-70329] Set the 'Maximum Shadows on Screen' limit to 65536.
Jira: [UUM-70329](https://jira.unity3d.com/browse/UUM-70329)

As reported in JIRA, rendering breaks when changing the `MaximumShadows on Screen` to a big number. This issue happens because the compute buffer size exceeds `2^31` bytes so it fails to allocate a buffer. By clamping the compute buffer size, prevent the allocation failure so the issue no longer happens.

The length of `_HDShadowDatas` is related to the number of lights. Currently, HDRP does have a light limit as below image so I set the `MaximumShadows on Screen` limit to `65536(= 2^16)` which is greater than the total light limit. So the maximum size of the `_HDShadowDatas` buffer will be `15MB(65536 * 240 bytes)`.

![image](https://media.github.cds.internal.unity3d.com/user/3842/files/1ebb3660-ab64-4aff-aead-551730952a40)

Added a clamping logic to the HDRP UI as well so users can recognize `65536` is the limitation.