[ShaderGraph] [UUM-76021] Fix node previews toggle shortcut
Jira: [UUM-76021](https://jira.unity3d.com/browse/UUM-76021)

As reported in JIRA, the `Collapse All Previews` and `Expand All Previews` options in the context menu work, but the keyboard shortcut does not. Since the name of the context menu option also includes the keyboard shortcut, I modified the shortcut API to make it compatible with all node previews when none are selected.

The behavior is:
1. If there's a selected node, the shortcut toggles the selected node previews as before.
2. If none are selected, the shortcut toggles all node previews depending on their `previewExpanded` status.