[ShaderGraph] [UUM-75921] Add sticky note checks to the group shortcuts
Jira: [UUM-75921](https://jira.unity3d.com/browse/UUM-75921)

As reported in JIRA, the Group/Ungroup Selection keyboard shortcuts do not work while the `Group Selection` and `Ungroup Selection` buttons in the context menu work. This happens when only sticky notes are selected. It does not occur when a sticky note is selected with at least one AbstractMaterialNode.

By adding the sticky note check to the keyboard shortcut, the shortcut now includes the sticky notes so the issue does not happen anymore.