Fix DefaultVolumeProfile UI errors
This PR fixes https://jira.unity3d.com/browse/UUM-77758

First (ad9e20cad58658fb5ff154cfa643e10a5f92caca) There's a fundamental problem with the way PropertyDrawers are accessed by RenderPipelineGraphicsSettingsContextMenuManager. Accessing them by calling `ScriptAttributeUtility.GetHandler(property).propertyDrawer` is not safe because there is no guarantee you will get a cached & valid drawer instance - the cache is cleared very frequently and in this case the PropertyDrawer is not the same instance as you might expect. For us it's a problem because DefaultVolumeProfileSettingsPropertyDrawer stores an Editor that the context menu wants to access. The bug was caused by the Editor instance being null because the PropertyDrawer instance was a different one. This has been solved by keeping the DefaultVolumeProfileEditor reference around in a static variable (we know this is fine because only one volume profile is the default).

Second (acf4baaa90a86c5f25c89c293b590354ec68a87a), when pipeline type changes, we need this PropertyDrawer to be rebuilt, because otherwise you will see URP volume editors that you're not supposed to see. 

Finally third (523fc9df89ed98d643e56bc906dd99af39afc049), we are including the RP type in the hash that prevents the GraphicsSettings UI from being refreshed. This way recreating the active RP won't cause UI to be refreshed, but changing to a different RP type will.