Prevent URP creation from UniversalCameraAdditionalData.OnDestroy
Fix error in Dedicated Server [UUM-75237](https://jira.unity3d.com/browse/UUM-75237)
What happened is that when destroying a UniversalCameraAdditionalData, in order to clear the buffer, the code was actually constructing the URP pipeline in case the pipeline was null.

Fix do the same without trying to construct the pipeline if it is not constructed (aka multiple pipeline scenario or headless server).