[UUM-75935] Prevent excessively large numeric errors when calculating area lighting.
Jira : [UUM-75935](https://jira.unity3d.com/browse/UUM-75935)

Cause: 
The reflection artifact described in the ticket occurs due to a numerical error in the Area Light computation. 
When calculating the Area Light, the function ComputeEdgeFactor(V1, V2) is used. 
If the two vectors V1 and V2 are collinear, the rcpLen value within the ComputeEdgeFactor function becomes abnormally large. 
Furthermore, if V1 and V2 are collinear, their dot product is negative, which causes the rcpLen value to affect the return value.
Consequently, this leads to the result value F becoming abnormally large.

Solution: 
To prevent this issue, the value of FLT_MIN within the max function, which ensures rcpLen is always finite, should be increased to FLT_EPS. 
This adjustment resolves the issue.