[Port] [2021.3] SG integration: Read unexposed global properties from cbuffer
Shader Graph allows to declare properties in a global context. These properties can be set from script, globally to all shaders at the same time. (see [Shader.SetGlobalVector](https://docs.unity3d.com/ScriptReference/Shader.SetGlobalVector.html for instance)).
These global properties work fine when used with a regular ShaderGraph material, but not with VFX outputs that use ShaderGraph.
The reason is that we were grouping them in a structure with the other exposed properties, when these should be read directly from the constant buffer. Just changing the declaration in the shader makes it work fine.