[Port] [2022.3] fix issue with EndCameraRendering being called before renderContext s…
This is a manual backport of https://github.cds.internal.unity3d.com/unity/unity/pull/52682. There was a small merge conflict because the `EndCameraRendering()` callback is called inside an `if` condition in Unity 6.0, but there's no such condition in 2022.3. The intent of the original PR is to move the callback down, so that it is called after the render context is submitted. I have retained the spirit of that change in this port.