[UUM-64808] Add alpha clamping to BloomPrefilter.compute
Jira: [UUM-64808](https://jira.unity3d.com/browse/UUM-64808)

As reported in JIRA, the bloom affects an entire mesh area even if the sampled texture color is black.
This issue happens when `ENABLE_ALPHA` keyword is enabled, the alpha channel in the color target could be greater than 1.0 after `Additive` blend calculations before the post-processing stage. It does not happen if the texture uses an alpha channel, and I think this is the expected setting. However, alpha clamping in the bloom prefilter seems necessary to handle an unexpected input and it is a more intuitive behavior for users.

![UUM-64808](https://media.github.cds.internal.unity3d.com/user/3842/files/422b8f7a-91d2-47c0-9cff-f29da06a8c69)