UUM-79124: Fix APV compatability with asset bundles and addressables.
This is a bugfix for an issue that causes APV to not work at all with scenes contained in Asset Bundles and Addressables. 

Before I explain the fix, I'll briefly explain the root cause: APV has 2 paths for including baked lighting data into Player builds: Via regular asset references, or via the [StreamingAssets](https://docs.unity3d.com/Manual/StreamingAssets.html) mechanism. The StreamingAssets path is used to support disk streaming, where the system needs full control over how files are read from disk. When using the StreamingAssets path, we don't want the data to be included into builds via regular asset references, as this would result in duplicate data in the build. For this reason, APV keeps the relevant asset references `null` until it is time to build the player. During build, we determine which of the 2 paths we are using, and either leave the asset references `null` and move the data to `StreamingAssets`, or populate the asset references and don't move the data to `StreamingAssets`. This setup works fine for Player builds, but the relevant build script _only_ runs for Player builds, and not for Asset Bundle builds.

My fix is to adapt the build script to also run for Asset Bundle builds. When we are processing a scene that will go into an Asset Bundle, we check if are using the regular asset reference path (which is required for Asset Bundle support), and update the asset references accordingly. This results in the previously missing lighting data being included into the bundle.

I had to expose a currently internal property in the content build pipeline bindings to implement this properly. I asked the relevant team and they were fine with it.