[WebGPU] Fallback for rendering layers 16 bit mask
https://jira.unity3d.com/browse/PLATGRAPH-3700

WebGPU does not support R16_Unorm as a render target format. 16 bit masks for rendering layers in URP use a R16_Unorm format as the render texture format, so we need a fallback for webgpu. This changes makes it fallback to r32_sfloat when the active graphics is webgpu and a 16 bit mask is set for the rendering layer feature.