Fix a APV cannot be enabled when scene and game view are not shown
Jira : [UUM-76078](https://jira.unity3d.com/browse/UUM-76078)

Cause : 
Changes to the Renderpipeline asset will cause the ProbeReferenceVolume to be initialize.
The above process only works when the actual render process is called at least once.
The problem is caused by the ProbeReferenceVolume not being initialize because the render process is not called when the Scene Viewer is not open.



Solution :
The contents of the Inspector in the ProbeVolumeEditor where the problem is occurring should be decoupled from the state of the ProbeReferenceVolume.
In other words, the behavior of the RenderPipelineAsset's ProbeVolume setting value should be decoupled from the behavior of the render not running because the Scene or Game viewer is not currently open.
So we changed the conditional checking logic to only check with the current ProbeVolume setting value of the RenderPipelineAsset. 


Memos :
Also, this issue happens exactly the same in HDRP, so I worked to make the same fix work in HDRP as well.