Added documentation for [Scene Color Sampling fails when using a cust…
Cause :
https://docs.unity3d.com/Packages/com.unity.render-pipelines.high-definition@17.0/manual/custom-post-processing-create-apply.html
https://docs.unity3d.com/Packages/com.unity.render-pipelines.high-definition@17.0/manual/custom-post-processing-scripts.html
If you create a CustomPostProcessEffect using the documentation in the link above, it will not automatically handle the _injectionPoint like it does when using CustomPass.
So, HDSceneColor node does not work correctly.

Solution :
This is not an issue that needs to be fixed.
If you end up using the same approach as above, you can use the HDSceneColourBuffer node.
![ure_007](https://media.github.cds.internal.unity3d.com/user/6528/files/6ec81d1f-56b3-4542-9c63-96367729b4c4)


We've added that to the documentation to avoid confusion for customers.

Memos :
The additional part of the document looks like this.

![ure_006](https://media.github.cds.internal.unity3d.com/user/6528/files/6395c87c-a94b-4c21-abb6-b89d6b0c9110)