Additional error checking when Importing textures into RenderGraph
URP RenderGraph [Jira](https://jira.unity3d.com/browse/URP-1297).

We assume in our RenderGraph code that an imported RenderTexture only has a single resource, depth OR color, but not both. 

In RenderGraph.ImportTexture, we describe the resource with a RenderTargetInfo that only has a single format, from the assumption that it describes a single resource. Therefor, importing a RenderTexture with both depth and color is invalid. 

This PR adds error handling to ensure this assumption is correct. This can break code that accidentally worked since the existing code picks either color or depth, in an undefined way (seems to prioritize color over depth). This seems to work but can lead to all kinds of hard to find bugs.

User projects can be affected by this change. An exception will now be thrown when a user incorrectly imports a RenderTexture with both color and depth.

This PR also fixes a number of bugs that this check discovered in URP and HDRP, leading to a reduction of GPU memory usage.